Moving the parsing code to a utility file
