+++ pages/benches/bench001.vue
@@ -103,64 +103,7 @@
       return extrudedMesh;
     }
 
-    // Function to parse path data and convert it into an array of points
-    function parsePathData(pathData) {
-      // Regular expression pattern to find all commands and their coordinates
-      const regex = /([mlhvcsqtaz])([^mlhvcsqtaz]*)/gi;
-
-      // Array to store the points
-      const points = [];
-
-      // Helper function to update the current point based on the command and its parameters
-      function updateCurrentPoint(command, parameters) {
-        const coordinates = parameters.split(/[ ,]/).map(parseFloat);
-
-        let index = 0;
-        while (index < coordinates.length) {
-          const dx = coordinates[index];
-          const dy = coordinates[index + 1];
-          currentPoint.x += dx;
-          currentPoint.y += dy;
-          points.push(new Vector3(currentPoint.x, currentPoint.y, 0));
-          index += 2; // Move to the next set of coordinates (x, y)
-        }
-      }
-
-      let currentPoint = new Vector3(0, 0, 0); // Starting point
-
-      let match;
-      while ((match = regex.exec(pathData)) !== null) {
-        const command = match[1];
-        const parameters = match[2].trim();
-        updateCurrentPoint(command, parameters);
-      }
-
-      return points;
-    }
-
-    // Create an array to store the objects
-    const pathsArray = [];
-
-    // Parse the SVG data to extract the path elements
-    const parser = new DOMParser();
-    const svgDocument = parser.parseFromString(svg, "image/svg+xml");
-    const pathElements = svgDocument.querySelectorAll("path");
-
-    // Iterate through each path and convert it to an object
-    pathElements.forEach((pathElement) => {
-      // Get the 'id' attribute of the path
-      const id = pathElement.getAttribute("id");
-
-      // Get the 'd' attribute of the path, which contains the path data
-      const pathData = pathElement.getAttribute("d");
-      const points = parsePathData(pathData);
-      const pathObject = {
-        id,
-        points
-      };
-
-      pathsArray.push(pathObject);
-    });
+    const pathsArray = extractSVG_babylon(svg);
 
     // Create a group to hold all the extruded paths
     const extrudedPathsGroup = new Mesh("extrudedPathsGroup", scene);
